Restrict 'added_on' field to have dump_only property
Changed 'added_on' field to have a 'dump_only' property in 'schema.py'. This is to ensure that the 'added_on' field cannot be manually modified by users and is automatically populated with the current datetime upon creation of a new category. By restricting 'added_on' to be dump_only, data consistency will be ensured.