Update .gitignore file
Added more files and directories to the .gitignore file. These include various IDE configuration files, a virtual environment directory, a local SQLite database file, and some specific error and notebook files. This is to ensure these files are not tracked as they are machine/user-specific configurations and binary db files.