Update relationship attribute in Product model
Reviewing the relationship established in the 'Product' model, the 'lazy' dynamic option for the 'image' attribute was removed and the 'backref' value was changed from 'product' to 'products'. This change improves data mapping efficiency between 'Product' and 'ProductImage' models and clarifies that a single image can be linked to multiple products.