Fix image opening by using stream attribute of image_file
Changed the method of image file opening in database/schema.py by reading the stream attribute of 'image_file'. Previously, the direct use of the 'image_file' object caused issues as the file was not being read correctly. Reading from the stream attribute ensures we have access to the file data as it's being uploaded, fixing the image processing.