"Expanded image upload rights to managers
Updated the imageAPI to allow not just admin users but also users with 'manager' role to upload and delete images. This accomplishes better role delegation in user management. Also, a new feature is added to return the image URL upon successful image upload to provide immediate confirmation for successful storage and retrieval."