"Add category deletion request function
Created a function in the managerAPI that lets a manager request to delete a category. It includes checks to ensure the requesting user is a manager, the category exists, it is not categorized as 'uncategorized', it does not have a pending deletion request and the requester originally added the category. This helps in better authorization and integrity of category deletion."