"Add user details retrieval function to login API
A new function 'get_user' was added to the loginAPI.py. It uses JWT for authentication and allows the authenticated users to retrieve their details stored in the User table. UserSchema from database.schema is also imported to facilitate the serialization of user data. This feature was added to provide users a way to view their account information."